Modifies initial check within handleCheck method
Added a check to make sure that the lastChecked variable was defined before entering the forEach loop on line 113. This check ensures that at least one checkbox was clicked beforehand, which prevents the auto-selection of every checkbox below the first value. This typically occurred when the user was holding shift and selecting their first checkbox, which caused all checkboxes below to display as checked. However, now the selection will only be filled if there are actually two checkboxes clicked beforehand.